+++ crates/ruma-state-res/src/event_auth/room_member/tests.rs
@@ -1,4 +1,7 @@
-use ruma_common::{third_party_invite::IdentityServerBase64PublicKey, Signatures};
+use ruma_common::{
+    room_version_rules::RoomVersionRules, third_party_invite::IdentityServerBase64PublicKey,
+    Signatures,
+};
 use ruma_events::{
     room::{
         join_rules::{JoinRule, Restricted, RoomJoinRulesEventContent},
@@ -17,7 +20,6 @@ use crate::{
         member_content_join, room_third_party_invite, to_pdu_event, zara, TestStateMap,
         INITIAL_EVENTS, INITIAL_EVENTS_CREATE_ROOM,
     },
-    RoomVersion,
 };
 
 #[test]
@@ -42,7 +44,7 @@ fn missing_state_key() {
     // Event should have a state key.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -71,7 +73,7 @@ fn missing_membership() {
     // Content should at least include `membership`.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -100,7 +102,7 @@ fn join_after_create_creator_match() {
     // Before v11, the `creator` of `m.room.create` must be the same as the state key.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -129,7 +131,7 @@ fn join_after_create_creator_mismatch() {
     // Before v11, the `creator` of `m.room.create` must be the same as the state key.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -158,7 +160,7 @@ fn join_after_create_sender_match() {
     // Since v11, the `sender` of `m.room.create` must be the same as the state key.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V11,
+        &RoomVersionRules::V11,
         room_create_event,
         fetch_state,
     )
@@ -187,7 +189,7 @@ fn join_after_create_sender_mismatch() {
     // Since v11, the `sender` of `m.room.create` must be the same as the state key.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V11,
+        &RoomVersionRules::V11,
         room_create_event,
         fetch_state,
     )
@@ -216,7 +218,7 @@ fn join_sender_state_key_mismatch() {
     // For join events, the sender must be the same as the state key.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -255,7 +257,7 @@ fn join_banned() {
     // A user cannot join if they are banned.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -294,7 +296,7 @@ fn join_invite_join_rule_already_joined() {
     // A user can send a join event in a room with `invite` join rule if they already joined.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -343,7 +345,7 @@ fn join_knock_join_rule_already_invited() {
     // invited.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V7,
+        &RoomVersionRules::V7,
         room_create_event,
         fetch_state,
     )
@@ -384,7 +386,7 @@ fn join_knock_join_rule_not_supported() {
     // for coverage.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -428,7 +430,7 @@ fn join_restricted_join_rule_not_supported() {
     // for coverage.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -472,7 +474,7 @@ fn join_knock_restricted_join_rule_not_supported() {
     // this is good for coverage.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -515,7 +517,7 @@ fn join_restricted_join_rule_already_joined() {
     // joined.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V8,
+        &RoomVersionRules::V8,
         room_create_event,
         fetch_state,
     )
@@ -567,7 +569,7 @@ fn join_knock_restricted_join_rule_already_invited() {
     // were invited.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V10,
+        &RoomVersionRules::V10,
         room_create_event,
         fetch_state,
     )
@@ -610,7 +612,7 @@ fn join_restricted_join_rule_missing_join_authorised_via_users_server() {
     // `join_authorised_via_users_server` property.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V8,
+        &RoomVersionRules::V8,
         room_create_event,
         fetch_state,
     )
@@ -656,7 +658,7 @@ fn join_restricted_join_rule_authorised_via_user_not_in_room() {
     // authorized by a user not in the room.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V8,
+        &RoomVersionRules::V8,
         room_create_event,
         fetch_state,
     )
@@ -711,7 +713,7 @@ fn join_restricted_join_rule_authorised_via_user_with_not_enough_power() {
     // authorized by a user with not enough power.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V8,
+        &RoomVersionRules::V8,
         room_create_event,
         fetch_state,
     )
@@ -787,7 +789,7 @@ fn join_restricted_join_rule_authorised_via_user() {
         // authorized by a user with enough power.
         check_room_member(
             RoomMemberEvent::new(&incoming_event),
-            &RoomVersion::V9,
+            &RoomVersionRules::V9,
             room_create_event,
             fetch_state,
         )
@@ -837,7 +839,7 @@ fn join_public_join_rule() {
         // A user can join a room with a `public` join rule.
         check_room_member(
             RoomMemberEvent::new(incoming_event),
-            &RoomVersion::V9,
+            &RoomVersionRules::V9,
             room_create_event.clone(),
             fetch_state,
         )
@@ -886,7 +888,7 @@ fn invite_via_third_party_invite_banned() {
     // A user cannot be invited via third party invite if they were banned.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -922,7 +924,7 @@ fn invite_via_third_party_invite_missing_signed() {
     // Third party invite content must have a `joined` property.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -961,7 +963,7 @@ fn invite_via_third_party_invite_missing_mxid() {
     // Third party invite content must have a `joined.mxid` property.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1000,7 +1002,7 @@ fn invite_via_third_party_invite_missing_token() {
     // Third party invite content must have a `joined.token` property.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1035,7 +1037,7 @@ fn invite_via_third_party_invite_mxid_mismatch() {
     // `mxid` of third party invite must match state key.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1088,7 +1090,7 @@ fn invite_via_third_party_invite_missing_room_third_party_invite() {
     // There must be an `m.room.third_party_invite` event with the same token in the state.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1125,7 +1127,7 @@ fn invite_via_third_party_invite_room_third_party_invite_sender_mismatch() {
     // `mxid` of third party invite must match state key.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1168,7 +1170,7 @@ fn invite_via_third_party_invite_with_room_missing_signatures() {
     // `signed` must have a `signatures` field.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1212,7 +1214,7 @@ fn invite_via_third_party_invite_with_room_empty_signatures() {
     // There is no signature to verify, we need at least one.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1260,7 +1262,7 @@ fn invite_via_third_party_invite_with_wrong_signature() {
     // No public key will manage to verify the signature.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1308,7 +1310,7 @@ fn invite_via_third_party_invite_with_wrong_signing_algorithm() {
     // Can't verify a signature with an unsupported algorithm, so there is no signature to verify.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1361,7 +1363,7 @@ fn invite_via_third_party_invite() {
     // Valid third party invite works.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1390,7 +1392,7 @@ fn invite_sender_not_joined() {
     // The sender of the invite must have joined the room.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1429,7 +1431,7 @@ fn invite_banned() {
     // The sender of the invite must have joined the room.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1458,7 +1460,7 @@ fn invite_already_joined() {
     // The sender of the invite must have joined the room.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1497,7 +1499,7 @@ fn invite_sender_not_enough_power() {
     // The sender must have enough power to invite in the room.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1526,7 +1528,7 @@ fn invite() {
     // The invite is valid.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1555,7 +1557,7 @@ fn leave_after_leave() {
     // User can only leave after `invite`, `join` or `knock`.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1584,7 +1586,7 @@ fn leave_after_join() {
     // User can leave after join.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1623,7 +1625,7 @@ fn leave_after_invite() {
     // User can leave after invite.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1662,7 +1664,7 @@ fn leave_after_knock() {
     // User can leave after knock.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V9,
+        &RoomVersionRules::V9,
         room_create_event,
         fetch_state,
     )
@@ -1702,7 +1704,7 @@ fn leave_after_knock_not_supported() {
     // membership if it's not supported by the room version, but this is good for coverage.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -1731,7 +1733,7 @@ fn leave_kick_sender_left() {
     // User can't kick if not joined.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -1770,7 +1772,7 @@ fn leave_unban_not_enough_power() {
     // User can't unban if not enough power.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -1809,7 +1811,7 @@ fn leave_unban() {
     // User can unban with enough power.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -1838,7 +1840,7 @@ fn leave_kick_not_enough_power() {
     // User can't kick if not enough power for it.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -1884,7 +1886,7 @@ fn leave_kick_greater_power() {
     // Can't kick user with greater power level.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -1930,7 +1932,7 @@ fn leave_kick_same_power() {
     // Can't kick user with same power level.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -1959,7 +1961,7 @@ fn leave_kick() {
     // Can kick user with enough power.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -1988,7 +1990,7 @@ fn ban_sender_not_joined() {
     // Can't ban user if not in room.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -2017,7 +2019,7 @@ fn ban_not_enough_power() {
     // Can't ban user if not enough power.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -2063,7 +2065,7 @@ fn ban_greater_power() {
     // Can't ban user with greater power level.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -2109,7 +2111,7 @@ fn ban_same_power() {
     // Can't ban user with same power level.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -2138,7 +2140,7 @@ fn ban() {
     // Can ban user with enough power.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V6,
+        &RoomVersionRules::V6,
         room_create_event,
         fetch_state,
     )
@@ -2167,7 +2169,7 @@ fn knock_public_join_rule() {
     // User can't knock if join rule is not `knock` or `knock_restricted`.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V11,
+        &RoomVersionRules::V11,
         room_create_event,
         fetch_state,
     )
@@ -2206,7 +2208,7 @@ fn knock_knock_join_rule() {
     // User can knock if room version supports it.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V7,
+        &RoomVersionRules::V7,
         room_create_event,
         fetch_state,
     )
@@ -2245,7 +2247,7 @@ fn knock_knock_join_rule_not_supported() {
     // User CANNOT knock if room version doesn't support it.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V5,
+        &RoomVersionRules::V5,
         room_create_event,
         fetch_state,
     )
@@ -2287,7 +2289,7 @@ fn knock_knock_restricted_join_rule() {
     // User can knock if room version supports it.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V10,
+        &RoomVersionRules::V10,
         room_create_event,
         fetch_state,
     )
@@ -2329,7 +2331,7 @@ fn knock_knock_restricted_join_rule_not_supported() {
     // User CANNOT knock if room version doesn't support it.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V5,
+        &RoomVersionRules::V5,
         room_create_event,
         fetch_state,
     )
@@ -2368,7 +2370,7 @@ fn knock_sender_state_key_mismatch() {
     // User cannot knock if state key doesn't match sender.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V7,
+        &RoomVersionRules::V7,
         room_create_event,
         fetch_state,
     )
@@ -2416,7 +2418,7 @@ fn knock_after_ban() {
     // User cannot knock if banned.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V7,
+        &RoomVersionRules::V7,
         room_create_event,
         fetch_state,
     )
@@ -2464,7 +2466,7 @@ fn knock_after_invite() {
     // User cannot knock after being invited.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V7,
+        &RoomVersionRules::V7,
         room_create_event,
         fetch_state,
     )
@@ -2503,7 +2505,7 @@ fn knock_after_join() {
     // User cannot knock after being invited.
     check_room_member(
         RoomMemberEvent::new(incoming_event),
-        &RoomVersion::V7,
+        &RoomVersionRules::V7,
         room_create_event,
         fetch_state,
     )