+++ crates/ruma-state-res/tests/it/resolve.rs
@@ -9,11 +9,11 @@ use std::{
 };
 
 use ruma_common::{
-    MilliSecondsSinceUnixEpoch, OwnedEventId, OwnedRoomId, OwnedUserId, RoomId, RoomVersionId,
-    UserId,
+    room_version_rules::RoomVersionRules, MilliSecondsSinceUnixEpoch, OwnedEventId, OwnedRoomId,
+    OwnedUserId, RoomId, RoomVersionId, UserId,
 };
 use ruma_events::{StateEventType, TimelineEventType};
-use ruma_state_res::{resolve, Event, RoomVersion, StateMap};
+use ruma_state_res::{resolve, Event, StateMap};
 use serde::{Deserialize, Serialize};
 use serde_json::{
     from_str as from_json_str, to_string_pretty as to_json_string_pretty,
@@ -200,15 +200,14 @@ fn test_resolve(paths: &[&str]) -> Snapshots {
             .expect("the m.room.create PDU's content should be valid")
             .room_version
     };
-    let room_version =
-        RoomVersion::new(&room_version_id).expect("room version should be supported");
+    let rules = room_version_id.rules().expect("room version should be supported");
 
     // Resolve PDUs in batches by file
     let mut pdus_by_id = HashMap::new();
     let mut batched_resolved_state = None;
     for pdus in &pdu_batches {
         batched_resolved_state = Some(
-            resolve_batch(&room_version, pdus, &mut pdus_by_id, &mut batched_resolved_state)
+            resolve_batch(&rules, pdus, &mut pdus_by_id, &mut batched_resolved_state)
                 .expect("batched state resolution step should succeed"),
         );
     }
@@ -217,7 +216,7 @@ fn test_resolve(paths: &[&str]) -> Snapshots {
 
     // Resolve all PDUs in a single step
     let atomic_resolved_state = resolve_batch(
-        &room_version,
+        &rules,
         pdu_batches.iter().flat_map(|x| x.iter()),
         &mut HashMap::new(),
         &mut None,
@@ -268,7 +267,7 @@ fn test_resolve(paths: &[&str]) -> Snapshots {
 ///
 /// # Arguments
 ///
-/// * `room_version`: The room's version.
+/// * `rules`: The rules of the room version.
 /// * `pdus`: An iterator of [`Pdu`]s to resolve, either alone or against the `prev_state`.
 /// * `pdus_by_id`: A map of [`OwnedEventId`]s to the [`Pdu`] with that ID.
 ///   * Should be empty for the first call.
@@ -277,7 +276,7 @@ fn test_resolve(paths: &[&str]) -> Snapshots {
 ///   * Should be `None` for the first call.
 ///   * Should not be mutated outside of this function.
 fn resolve_batch<'a, I, II>(
-    room_version: &RoomVersion,
+    rules: &RoomVersionRules,
     pdus: II,
     pdus_by_id: &mut HashMap<OwnedEventId, Pdu>,
     prev_state: &mut Option<StateMap<OwnedEventId>>,
@@ -310,8 +309,7 @@ where
         auth_chain_sets.push(auth_events_dfs(&*pdus_by_id, pdu)?);
     }
 
-    resolve(room_version, &state_sets, auth_chain_sets, |x| pdus_by_id.get(x).cloned())
-        .map_err(Into::into)
+    resolve(rules, &state_sets, auth_chain_sets, |x| pdus_by_id.get(x).cloned()).map_err(Into::into)
 }
 
 /// Depth-first search for the `auth_events` of the given PDU.